Upgrade to Neo4j Java Driver 6.0 (compile)
This change only introduces the bump of the driver version to 6.0.
As a consequence the transitive driver-slim dependency of caniuse was
excluded to only have one driver on the classpath.
The `AuthTokenManager` currently does not know anything about Keycloak,
because the retrying driver is not in use anymore but only the reauthn
lib.

One thing that came up during migrating the test setup is that we dont
`consume` all writes/setup invocation. This could lead to the problem
that we still have open cursors. We should spend some time and check all
tests calling driver's `executeWrite` or similar for missing `consume`
calls.