feat: introduce byte, short, decimal, timestampNTZ types (#816)
adds support to natively write these spark types to neo4j:

- Byte: will convert to Integer, therefore cannot read-back as Byte
- Short: will convert to Integer, therefore cannot read-back as Short
- Decimal: will convert to String, therefore cannot read-back as Decimal
- Binary: will convert to ByteArray and will be read-back as ByteArray
- TimestampNTZType: will convert to LocalDateTime

BREAKING CHANGE:

- already existing Timestamp now converts to ZonedDateTime, not LocalDateTime.

Before, writing a Timestamp would remove the timezone information.
Now it will be normalized to UTC-time instead. If you want to retain old behaviour,
please use TimestampNTZType or be mindful when you read the stamp.

Assuming you interpret time zone when reading and writing, you should not notice.
But if you indeed intend to use LocalDateTime, this is moved and maps to TimestampNTZType